# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory API Documentation",
        "description": "API documentation for the Inventory project",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Inventory API Server"
        }
    ],
    "paths": {
...
        "/api/poles": {
            "get": {
                "tags": [
                    "Poles"
                ],
                "summary": "Lister tous les pôles",
                "operationId": "099822f0589663d8d415a7a1510a5325",
                "responses": {
                    "200": {
                        "description": "Liste des pôles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_pole": {
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "nom": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Poles"
                ],
                "summary": "Créer un pôle",
                "operationId": "04f113bfa4dc44c1ac75e4510b1e9e03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Pôle Logistique"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Gère les opérations logistiques"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pôle créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_pole": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "nom": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/poles/{id}": {
            "get": {
                "tags": [
                    "Poles"
                ],
                "summary": "Afficher un pôle",
                "operationId": "4870b878a562e29c8e51cfd1e2de03e8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du pôle",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pôle trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_pole": {
                                            "type": "string"
                                        },
                                        "nom": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pôle introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Poles"
                ],
                "summary": "Modifier un pôle",
                "operationId": "1ef35bd16e1f43c591ec75016ad0deed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du pôle",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pôle mis à jour"
                    },
                    "404": {
                        "description": "Pôle introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Poles"
                ],
                "summary": "Supprimer un pôle",
                "operationId": "32b9f28dcd2d6db39d6300537d625e67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pôle supprimé"
                    },
                    "404": {
                        "description": "Pôle introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}